requirements.txt
streamlit 
openai==0.28
replicate 
python-dotenv 
requests 
pillow